CHANGE: class="label" to class="set-label"
Breaking old behaviour, but hopefully with good reason.

BS3 already have a .label class which formats text different to the form labels.

When using sets, or ':as => :radio' the output adds a <span class="label">... tag which does not render well in the demo-site and when using BS3.

Therefore I changed the class from .label to .set-label.

If this can be achived in a non-breaking way, please do so.