Three (3) failing tests & expected output
Three (3) failing tests remaining to resolve around :checkboxset/:radioset functionality.

A single :checkbox/:radio or a :checkboxset/:radioset should be wrapped in a
<div class="has-error">...</div> tag. (see example below)

Expected output format should be as follows:

    <!-- single input -->
    <div class="has-error">
      <div class="checkbox/radio">
        <label>
          <input ...>
        </label>
      </div>
      <span class="help-block with-errors">Message...</span>
    </div>

    <!-- set of inputs -->
    <div class="has-error">

      <div class="checkbox/radio">
        <label>
          <input ...>
        </label>
      </div>

      <div class="checkbox/radio">
        <label>
          <input ...>
        </label>
      </div>

      <span class="help-block with-errors">Message...</span>
    </div>

I have tried many hacks to try to resolve this issue, but without success.

The problem seems to be in the `formatter.rb` file around lines 250, where the error and error message is tagged onto the last input (checkbox/radio).

All tests have the correct expected output format, and if this issue can be resolved then all tests will pass and most of my work is done.