Added updated specs & BS3 HTML output reference
Please note!  The BS3 HTML reference document illustrates the expected output for which all tests are based upon.

The error handling for :checkboxset & :radioset is unclear, so I have added two alternative versions.

In my mind I believe the current Forme output (alternative A) is the correct output, but that output is not the same as the default error output for :checkbox/:radio inputs.

Feedback is most welcome!!